UHF-9708: Move Global annoucements client to helfi_etusivu_entities module
