+++ modules/helfi_etusivu_entities/README.md
@@ -1,0 +1,8 @@
+# Remote entities
+
+Remote entities module allows fetching announcements from `Etusivu`-instance.
+It utilizes `json-api` and `external_entities`-module to transfer the data between instances.
+
+## How to set up locally
+
+Local setup requires Etusivu-instance to be up and running with some relevant data created to it.
