+++ modules/helfi_etusivu_entities/config/install/external_entities.external_entity_type.helfi_announcements.yml
