+++ modules/helfi_etusivu_entities/helfi_etusivu_entities.services.yml
@@ -1,0 +1,6 @@
+services:
+  logger.channel.helfi_etusivu_entities:
+    parent: logger.channel_base
+    arguments: [ 'helfi_etusivu_entities' ]
+
+
