+++ modules/helfi_etusivu_entities/src/Plugin/ExternalEntities/StorageClient/EtusivuEntityBase.php
@@ -2,10 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Drupal\helfi_global_announcement\Plugin\ExternalEntities\StorageClient;
+namespace Drupal\helfi_etusivu_entities\Plugin\ExternalEntities\StorageClient;
 
 use Drupal\Core\Entity\EntityStorageException;
-use Drupal\Core\Language\LanguageInterface;
 use Drupal\Core\Language\LanguageManagerInterface;
 use Drupal\Core\Utility\Error;
 use Drupal\external_entities\ExternalEntityInterface;
@@ -17,26 +16,23 @@ use GuzzleHttp\Exception\GuzzleException;
 use GuzzleHttp\Exception\RequestException;
 use GuzzleHttp\Psr7\Query;
 use GuzzleHttp\Utils;
-use Psr\Log\LoggerInterface;
+use Psr\Log\LoggerAwareInterface;
+use Psr\Log\LoggerAwareTrait;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
- * External entity storage client for News feed entities.
- *
- * @ExternalEntityStorageClient(
- *   id = "helfi_announcements",
- *   label = @Translation("Helfi: Announcements"),
- *   description = @Translation("Retrieves announcements from helfi")
- * )
+ * Base class for etusivu external entity storage client.
  */
-final class Announcements extends ExternalEntityStorageClientBase {
+abstract class EtusivuEntityBase extends ExternalEntityStorageClientBase implements LoggerAwareInterface {
+
+  use LoggerAwareTrait;
 
   /**
-   * Custom cache tag for announcements.
+   * Custom cache tag.
    *
    * @var string
    */
-  public static string $customCacheTag = 'helfi_external_entity_announcement';
+  public static string $customCacheTag = 'helfi_external_entity';
 
   /**
    * The active endpoint environment.
@@ -46,13 +42,6 @@ final class Announcements extends ExternalEntityStorageClientBase {
   private ?Environment $environment = NULL;
 
   /**
-   * The current language service.
-   *
-   * @var \Drupal\Core\Language\LanguageManagerInterface
-   */
-  private LanguageManagerInterface $languageManager;
-
-  /**
    * The HTTP client service.
    *
    * @var \GuzzleHttp\ClientInterface
@@ -60,11 +49,11 @@ final class Announcements extends ExternalEntityStorageClientBase {
   private ClientInterface $client;
 
   /**
-   * The logger.
+   * The current language service.
    *
-   * @var \Psr\Log\LoggerInterface
+   * @var \Drupal\Core\Language\LanguageManagerInterface
    */
-  private LoggerInterface $logger;
+  protected LanguageManagerInterface $languageManager;
 
   /**
    * {@inheritdoc}
@@ -76,8 +65,8 @@ final class Announcements extends ExternalEntityStorageClientBase {
     $plugin_definition,
   ) : self {
     $instance = parent::create($container, $configuration, $plugin_id, $plugin_definition);
-    $instance->languageManager = $container->get('language_manager');
     $instance->client = $container->get('http_client');
+    $instance->languageManager = $container->get('language_manager');
 
     /** @var \Drupal\helfi_api_base\Environment\EnvironmentResolver $environmentResolver */
     $environmentResolver = $container->get('helfi_api_base.environment_resolver');
@@ -88,7 +77,7 @@ final class Announcements extends ExternalEntityStorageClientBase {
     }
     catch (\InvalidArgumentException) {
     }
-    $instance->logger = $container->get('logger.factory')->get('helfi_announcements');
+    $instance->setLogger($container->get('logger.channel.helfi_etusivu_entities'));
 
     return $instance;
   }
@@ -133,43 +122,10 @@ final class Announcements extends ExternalEntityStorageClientBase {
   }
 
   /**
-   * {@inheritdoc}
-   */
-  public function query(
-    array $parameters = [],
-    array $sorts = [],
-    $start = NULL,
-    $length = NULL,
-  ) : array {
-    $query = [
-      'fields[node--announcements]' => 'id',
-      'fields[status]' => 1,
-      'filter[status][value]' => 1,
-      'filter[field_publish_externally][value]' => 1,
-    ];
-
-    if ($start) {
-      $query['page[offset]'] = $start;
-    }
-
-    if ($length) {
-      $query['page[limit]'] = $length;
-    }
-
-    // If filter is missing language set it manually.
-    if (!isset($query['filter[langcode]'])) {
-      $language = $this->languageManager
-        ->getCurrentLanguage(LanguageInterface::TYPE_CONTENT)
-        ->getId();
-      $query['filter[langcode]'] = $language;
-    }
-
-    return $this->request($query, $query['filter[langcode]']);
-  }
-
-  /**
    * Creates a request against JSON:API.
    *
+   * @param string $endpoint
+   *   The jsonapi endpoint.
    * @param array $parameters
    *   The query parameters.
    * @param string $langcode
@@ -178,13 +134,14 @@ final class Announcements extends ExternalEntityStorageClientBase {
    * @return array
    *   An array of entities.
    */
-  private function request(array $parameters, string $langcode) : array {
+  protected function request(string $endpoint, array $parameters, string $langcode) : array {
     if (!$this->environment) {
       return [];
     }
     try {
-      $uri = vsprintf('%s/jsonapi/node/announcement?%s', [
+      $uri = vsprintf('%s/jsonapi%s?%s', [
         $this->environment->getInternalAddress($langcode),
+        $endpoint,
         Query::build($parameters),
       ]);
       $content = $this->client->request('GET', $uri);
