+++ modules/helfi_global_announcement/helfi_global_announcement.info.yml
@@ -4,3 +4,4 @@ core_version_requirement: '^9 || ^10'
 dependencies:
   - external_entities:external_entities
   - helfi_node_announcement:helfi_node_announcement
+  - helfi_platform_config:helfi_etusivu_entities
