+++ modules/helfi_global_announcement/helfi_global_announcement.module
@@ -8,10 +8,7 @@
 declare(strict_types=1);
 
 use Drupal\Core\Cache\Cache;
-use Drupal\Core\Entity\EntityTypeInterface;
-use Drupal\Core\StringTranslation\TranslatableMarkup;
-use Drupal\entity\BundleFieldDefinition;
-use Drupal\helfi_global_announcement\Plugin\ExternalEntities\StorageClient\Announcements;
+use Drupal\helfi_etusivu_entities\Plugin\ExternalEntities\StorageClient\Announcements;
 
 /**
  * Gets the block configurations.
@@ -60,39 +57,6 @@ function helfi_global_announcement_get_block_configurations(string $theme) : arr
       ],
     ],
   ];
-}
-
-/**
- * Implements hook_entity_bundle_field_info_alter().
- */
-function helfi_global_announcement_entity_bundle_field_info_alter(
-  &$fields,
-  EntityTypeInterface $entity_type,
-) : void {
-  if ($entity_type->id() === 'helfi_announcements') {
-
-    // Additional entity info fields.
-    $entity_info_fields = [
-      'published_at' => new TranslatableMarkup('Published at'),
-      'unpublish_on' => new TranslatableMarkup('Unpublish on'),
-      'notification' => new TranslatableMarkup('Notification'),
-      'langcode' => new TranslatableMarkup('Langcode'),
-      'body' => new TranslatableMarkup('Body'),
-      'status' => new TranslatableMarkup('Status'),
-      'announcement_type' => new TranslatableMarkup('Announcement type'),
-      'announcement_link_text' => new TranslatableMarkup('Announcement link text'),
-      'announcement_link_url' => new TranslatableMarkup('Announcement link url'),
-      'announcement_assistive_technology_close_button_title' => new TranslatableMarkup('Assistive technology close button title'),
-    ];
-
-    foreach ($entity_info_fields as $field_name => $field_label) {
-      $fields[$field_name] = BundleFieldDefinition::create('string')
-        ->setName($field_name)
-        ->setLabel($field_label)
-        ->setDisplayConfigurable('view', TRUE)
-        ->setDisplayConfigurable('form', TRUE);
-    }
-  }
 }
 
 /**
